requirements.txt
config
cv2==4.2.0
Werkzeug==1.0.0
requests==2.22.0
numpy==1.18.1
Flask==1.1.1
scikit_image==0.16.2
matplotlib==3.1.3
imutils==0.5.3
pdf2image==1.12.1
PyPDF2==1.26.0
pytesseract==0.3.3
python-docx==0.8.10